[#18] Parser defaults to returning 16 rows when a limit is not defined.
This protects the server from out-of-memory errors.

If a client wants more rows to be returned, all they need to do is
specify the number of rows in the query string.

The default can be overridden via the "options::default_number_of_rows".
"options::default_number_of_rows" is an unsigned 16-bit integer.

The limit embedded in the query string supersedes the value defined in
the "options" data type.