[#9] Parser uses correct type for CAST() when handling DATA_RESC_HIER.
The API plugin passes a database hint to the parser so that the parser
knows which target type to use for the DATA_RESC_HIER WITH-clause.

The API plugin uses the key defined in server_config.json at the
following location:

    - /plugin_configuration/database/<key>

Where <key> will be one of the following:

    - postgres
    - mysql
    - oracle

For those situations where <key> does not match one of those values, the
parser defaults to using the value tied to "postgres".

This commit also renames the CTE table to "cte_drh" to avoid any future
table naming collisions.