[#6] API plugin now supports federation.
If the plugin determines that it needs to redirect to a server in a
federated zone, the source server MUST be configured for federation.
That means, there must be an appropriate federation stanza defined in
server_config.json for the source server and the federated server.

Failing to meet this requirement will result in a
SYS_SIGNED_SID_NOT_MATCHED error.