[#26] Parser reorders tables to satisfy join expressions.
This is achieved by storing the index of an edge in the edge property
and using that index to lookup the edge definition.

The edge definition is a pair of indices which define the link between
tables. Each edge definition matches the expected order for the SQL join
expression. That is, the first value of the pair is the index of the
table that should be on the left-hand side of the expression. The second
value of the pair is the index of the table that should be on the
right-hand side of the expression.

By using the index stored in the edge property, we are able to reorder
the table aliases to meet the requirements of the join expression.